+++ cache-connector_test.go
@@ -11,7 +11,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/bborbe/teamvault-utils/v4"
+	"github.com/bborbe/teamvault-utils/v5"
 )
 
 var _ = Describe("CacheConnector", func() {