+++ cmd/teamvault-login/main.go
@@ -18,8 +18,8 @@ import (
 	libtime "github.com/bborbe/time"
 	"golang.org/x/term"
 
-	teamvault "github.com/bborbe/teamvault-utils/v4"
-	"github.com/bborbe/teamvault-utils/v4/factory"
+	teamvault "github.com/bborbe/teamvault-utils/v5"
+	"github.com/bborbe/teamvault-utils/v5/factory"
 )
 
 func main() {