+++ cmd/teamvault-url/main_test.go
@@ -15,7 +15,7 @@ import (
 var _ = Describe("Teamvault Url", func() {
 	It("Compiles", func() {
 		var err error
-		_, err = gexec.Build("github.com/bborbe/teamvault-utils/v4/cmd/teamvault-url")
+		_, err = gexec.Build("github.com/bborbe/teamvault-utils/v5/cmd/teamvault-url")
 		Expect(err).NotTo(HaveOccurred())
 	})
 })