+++ keychain_impl_integration_test.go
@@ -15,7 +15,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	teamvault "github.com/bborbe/teamvault-utils/v4"
+	teamvault "github.com/bborbe/teamvault-utils/v5"
 )
 
 var _ = Describe("DarwinKeychain integration", func() {